fix page fault crash: disable strict copy to/from userspace
in kernel 4.4 the semantic of how copy to/from user space is much stricter
when domain pan is enabled. So CONFIG_CPU_SW_DOMAIN_PAN shoud turn off to fix
the page fault crash.

Moreover, the way of splitting VM is back to default. As we decrease the amount
of GPU memory in device tree to the 256MB. In case of increasing the GPU memory,
maybe it is needed to adjust VMSPLIT.

Signed-off-by: Reyhaneh Yazdani <reyhane.y84@gmail.com>