wdt:dts: apply updated wtd-driver for gpio-wdt. Add wdt in dts
HW watchdog is designed in PPC board, which controls by gpio pins.
Current gpio-wdt driver does not work properly. It is updated to
the driver in kernel 4.12. In this driver, timeout is not defined
from user, because it is fixed by HW. Thus, unnecessary timer comcepts
is removed in the new driver.

correcspond watchdog nodei n device tree is also added.

Signed-off-by: reyhanehy <reyhane.y84@gmail.com>