dts: add spi1 port
SPi1 port is enabled. 2 GPIOs are set to work as chip select.
spi slave devices can be added as children of the spi bus node.

The Linux kernel offers a sample client driver called spidev that gives you
r/write access to the SPI bus through the /dev interface(CONFIG_SPI_SPIDEV).

In the default device tree spidev nodes are commented, because spidev is a
detail of how Linux controls a device rather than a description of the hardware
in the system. It means any SPI device could be a spidev.  To prevent any
mistake, the driver  generates a warning if spidev is instantiated as a DT
node without an ID in the match table.

Signed-off-by: reyhanehy <reyhane.y84@gmail.com>