pcie: back to internal clk.Define dts only for ext_osc
Since it would not be possible to have dynamic selection in dts about
internal and external clk, It is decided to use internal clk and
in case of request, use external clk.

To this point, new dts is defined which has all settings related to external clk.
It is just needed to include it.

Inside driver, all modification related to external is now depended on a definition
(EXT_OSC). If it is set to 1, then external clk codes are compiled.

This is also defined for clk-imx6q.c file.

All modifications are came from nxp community:

https://community.nxp.com/thread/445479
https://community.nxp.com/thread/455537

Signed-off-by: reyhanehy <reyhane.y84@gmail.com>