[media] imx-ipu-vout: use hardirq
Threaded IRQs have latencies up to 8ms on i.MX6 SMP,
which makes it much more likely to hit the IPU killswitch.

Signed-off-by: Lucas Stach <l.stach@pengutronix.de>
---
We might want to look into those latencies again to see
if we can make them a lot smaller, but still using a
hardirq for this hacky baseplane overlay is the right
thing to do.

Signed-off-by: reyhanehy <reyhane.y84@gmail.com>