imx-media: ovl: initialization of plane1 as overlay
In kernel 4.9, the plane1 of the IPU is initialized inside ipuv3-crtc.
Therefore, it does not need to assign resources again inside ovl driver.

The reources should only be assigned to pointer variables of ovl structure.

Note: for background it is needed to define a pointer variable which points to
plane0. Therefore two ipuv3-channel are defined. one for BG and the other for FG

Signed-off-by: reyhanehy <reyhane.y84@gmail.com>